+++ core-bundle/src/Doctrine/Schema/DcaSchemaProvider.php
@@ -102,6 +102,11 @@ class DcaSchemaProvider
                     $table->addOption('collate', $match[1]);
                 }
             }
+
+            // The default InnoDB row format before MySQL 5.7.9 is "Compact" but innodb_large_prefix requires "DYNAMIC"
+            if ($table->hasOption('engine') && 'InnoDB' === $table->getOption('engine')) {
+                $table->addOption('row_format', 'DYNAMIC');
+            }
         }
     }